// src/anime-info.interface.ts
export interface AnimeInfoInterface {
    extension?: string;
    releaseGroup?: string;
    resolution?: string;
    season?: string;
    episodeOrMovieNumber?: string;
    version?: string;
    videoTerm? :string;
    audioTerm?: string;
    checksum?: string;
    language?: string;
    source?: string;
    subtitles?: string;
    otherInfo?: string;
    volume?: string;
    year?: string;
    deviceCompatibility?: string;
    releaseInfo?: string;
    animeType1?: string;
    animeType2?: string;
    episodeTitle?: string;
    title?: string;
}